moving a few things
