+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/file/demographics/race/validation/ExistingRaceCategory.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.patient.profile.race.validate.category;
+package gov.cdc.nbs.patient.file.demographics.race.validation;
 
 public record ExistingRaceCategory(String identifier, String description) {
 }
