+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/profile/create/NewPatient.java
@@ -11,7 +11,7 @@ import gov.cdc.nbs.patient.demographics.identification.IdentificationDemographic
 import gov.cdc.nbs.patient.profile.mortality.MortalityDemographic;
 import gov.cdc.nbs.patient.demographics.name.NameDemographic;
 import gov.cdc.nbs.patient.demographics.phone.PhoneDemographic;
-import gov.cdc.nbs.patient.profile.race.RaceDemographic;
+import gov.cdc.nbs.patient.demographics.race.RaceDemographic;
 
 import java.util.List;
 import java.util.Optional;
