+++ apps/modernization-api/src/test/java/gov/cdc/nbs/patient/file/demographics/race/validation/ExistingRaceCategoryResultSetMapperTest.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.patient.profile.race.validate.category;
+package gov.cdc.nbs.patient.file.demographics.race.validation;
 
 import org.junit.jupiter.api.Test;
